tests/requirements.txt
invenio-app>=1.3.3
invenio-db[postgresql,mysql,versioning]>=1.0.14,<2.0.0
pytest-invenio>=1.4.11
invenio_search[opensearch2]>=2.0.0
Werkzeug>=2.2.3
Flask-Login>=0.6.1
pyyaml>=6.0
requests>=2.28.1